"create file structure of project fix"
