+++ database/database.go
@@ -22,7 +22,12 @@ func tableExists(db *sql.DB, tableName string) bool {
 
 func InitializeDatabase() (*sql.DB, error) {
 
-	dbFile := filepath.Join("database", "scheduler.db")
+	// basePath, err := os.Getwd()
+	// if err != nil {
+	// 	return nil, err
+	// }
+
+	dbFile := filepath.Join("../database", "scheduler.db")
 	//_, err = os.Stat(dbFile)
 
 	var install bool
