+++ internal/app/app.go
@@ -1,0 +1,9 @@
+package app
+
+import (
+	"database/sql"
+)
+
+type App struct {
+	DB *sql.DB
+}
