+++ internal/domain/services/nextDate/nextDate.go
@@ -2,7 +2,6 @@ package nextDate
 
 import (
 	"fmt"
-	"net/http"
 	"strconv"
 	"strings"
 	"time"
@@ -14,7 +13,7 @@ func NextDate(now time.Time, date string, repeat string) (string, error) {
 	var err error
 
 	if date == "" {
-		parsedDate = time.Now()
+		parsedDate = now
 	} else {
 		parsedDate, err = time.Parse(layout, date)
 		if err != nil {
@@ -25,7 +24,7 @@ func NextDate(now time.Time, date string, repeat string) (string, error) {
 		if parsedDate.After(now) {
 			return parsedDate.Format(layout), nil
 		}
-		return time.Now().Format(layout), nil
+		return now.Format(layout), nil
 
 	}
 	if repeat == "y" {
@@ -44,43 +43,25 @@ func NextDate(now time.Time, date string, repeat string) (string, error) {
 		days, err := strconv.Atoi(repeatDate[1])
 		if err != nil || days <= 0 || days > 400 {
 			return "", fmt.Errorf("неправильное количество дней: %v", err)
-		} else if days == 1 {
-			return parsedDate.Format(layout), nil
 		}
+		// if parsedDate.Equal(now) {
+		// 	return parsedDate.Format(layout), nil
+		// }
 
-		parsedDate = parsedDate.AddDate(0, 0, days)
-		for !parsedDate.After(now) {
+		//parsedDate = parsedDate.AddDate(0, 0, days)
+		// for !parsedDate.After(now) {
+		// 	parsedDate = parsedDate.AddDate(0, 0, days)
+		// }
+		// return parsedDate.Format(layout), nil
+
+		fmt.Printf("Начальная дата: %s, now: %s, шаг: %d\n", parsedDate.Format(layout), now.Format(layout), days)
+
+		for parsedDate.Before(now) || parsedDate.Equal(now) {
 			parsedDate = parsedDate.AddDate(0, 0, days)
+			fmt.Printf("Обновлённая дата: %s\n", parsedDate.Format(layout))
 		}
 		return parsedDate.Format(layout), nil
+
 	}
 	return "", fmt.Errorf("неподдерживаемое правило повторения: %s", repeat)
-}
-
-func HandleNextDate(w http.ResponseWriter, r *http.Request) {
-	layout := "20060102"
-
-	nowParam := r.URL.Query().Get("now")
-	dateParam := r.URL.Query().Get("date")
-	repeatParam := r.URL.Query().Get("repeat")
-
-	if nowParam == "" || dateParam == "" || repeatParam == "" {
-		http.Error(w, "Отсутствуют необходимые параметры", http.StatusBadRequest)
-		return
-	}
-
-	now, err := time.Parse(layout, nowParam)
-	if err != nil {
-		http.Error(w, "Неправильный формат даты", http.StatusBadRequest)
-		return
-	}
-
-	nextDate, err := NextDate(now, dateParam, repeatParam)
-	if err != nil {
-		http.Error(w, "Ошибка вычисления следующей даты", http.StatusBadRequest)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	fmt.Fprint(w, nextDate)
 }
